Added authorization requirements to prevent DB abuse

+++ main.py
@@ -60,19 +60,44 @@ def google_home():
 
 @app.route('/internal/scrape/generate_new_menu_data')
 def daily_scrape():
+    token = request.args.get('token')
+    permission = get_authorization(token)
+    if permission != 'admin' and permission != 'cron':
+        return 'NOT AUTHORIZED', 403
     return scrape.daily_scrape()
 
 
 @app.route('/internal/db/clear_cache')
 def clear_cache():
+    token = request.args.get('token')
+    permission = get_authorization(token)
+    if permission != 'admin' and permission != 'cron':
+        return 'NOT AUTHORIZED', 403
     return tracked_item.purge_old_cached_queries()
 
 
 @app.route('/internal/notify/today')
 def send_daily_messages():
+    token = request.args.get('token')
+    permission = get_authorization(token)
+    if permission != 'admin' and permission != 'cron':
+        return 'NOT AUTHORIZED', 403
+
     results = response.generate_daily_messages()
     for result in results:
         messenger.message_group(result[1], result[0])
+    return 'Notified {} groups'.format(len(results))
+
+
+def get_authorization(token):
+    if token is None:
+        return None
+    with db.db_pool.connect() as conn:
+        cur = conn.execute("SELECT permission FROM auth.users WHERE token = %s;", token)
+        rs = cur.fetchall()
+        if len(rs) == 0:
+            return None
+        return rs[0][0]
 
 
 if __name__ == '__main__':
